Move configuration into its own file.
